fyp/tests/tests/base_tests/base_test.py
#!/usr/bin/env python
# -*- coding:utf8 -*-

import os
import unittest
from appium import webdriver
from tests.pages import *
from tests.utils.appium_wrapper import startAppium, stopAppium


PATH = lambda p: os.path.abspath(
    os.path.join(os.path.dirname(__file__), p)
)


class BaseTest(unittest.TestCase):
    """Basis for all tests."""
    def setUp(self):

        startAppium()
        desired_caps = {}
        desired_caps['platformName'] = 'Android'
        desired_caps['platformVersion'] = os.environ['platformVersion']
        desired_caps['deviceName'] = '192.168.56.101:5555'
        desired_caps['app'] = os.environ['app']
        self.driver = webdriver.Remote('http://localhost:4723/wd/hub', desired_caps)

    def tearDown(self):
        """Shuts down the driver."""
        self.driver.quit()
        stopAppium()
